refactor: Piece Detail 패키지로 분리
