+++ src/main/java/chess/piece/Pawn.java
@@ -1,9 +1,12 @@
 package chess.piece;
 
-import static chess.piece.Color.BLACK;
-import static chess.piece.Color.WHITE;
+import static chess.piece.detail.Color.BLACK;
+import static chess.piece.detail.Color.WHITE;
 
 import chess.game.MoveCommand;
+import chess.piece.detail.Color;
+import chess.piece.detail.Direction;
+import chess.piece.detail.Name;
 import chess.position.Position;
 
 public class Pawn extends AbstractPiece {
