+++ src/main/java/chess/piece/Rook.java
@@ -1,6 +1,9 @@
 package chess.piece;
 
 import chess.game.MoveCommand;
+import chess.piece.detail.Color;
+import chess.piece.detail.Direction;
+import chess.piece.detail.Name;
 import chess.position.Position;
 import java.util.List;
