+++ src/main/java/chess/piece/detail/Color.java
@@ -1,4 +1,4 @@
-package chess.piece;
+package chess.piece.detail;
 
 public enum Color {
     BLACK, WHITE;
