+++ src/test/java/chess/piece/PawnTest.java
@@ -1,7 +1,7 @@
 package chess.piece;
 
-import static chess.piece.Color.BLACK;
-import static chess.piece.Color.WHITE;
+import static chess.piece.detail.Color.BLACK;
+import static chess.piece.detail.Color.WHITE;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import chess.game.MoveCommand;
